+++ player.h
@@ -52,7 +52,7 @@ struct displaytime {
 };
 
 long get_next_subsong(struct gbs *gbs);
-int get_prev_subsong(struct gbs *gbs);
+long get_prev_subsong(struct gbs *gbs);
 long nextsubsong_cb(struct gbs *gbs, void *priv);
 void update_displaytime(struct displaytime *time, const struct gbs_status *status);
 struct gbs *common_init(int argc, char **argv);
